+++ plugins/setup/src/main/java/com/wazuh/setup/index/StreamIndex.java
@@ -19,9 +19,8 @@ package com.wazuh.setup.index;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.opensearch.ResourceAlreadyExistsException;
-import org.opensearch.action.admin.indices.alias.Alias;
-import org.opensearch.action.admin.indices.create.CreateIndexRequest;
-import org.opensearch.action.admin.indices.create.CreateIndexResponse;
+import org.opensearch.action.admin.indices.datastream.CreateDataStreamAction;
+import org.opensearch.action.support.clustermanager.AcknowledgedResponse;
 
 import com.wazuh.setup.settings.PluginSettings;
 
@@ -32,58 +31,58 @@ import com.wazuh.setup.settings.PluginSettings;
 public class StreamIndex extends WazuhIndex {
     private static final Logger log = LogManager.getLogger(StreamIndex.class);
 
-    private final String alias;
-
     /**
      * Constructor.
      *
      * @param index index name.
      * @param template index template name.
-     * @param alias index alias name for advanced management such as automatic rollover.
      */
-    public StreamIndex(String index, String template, String alias) {
+    public StreamIndex(String index, String template) {
         super(index, template);
-        this.alias = alias;
     }
 
     /**
-     * Overrides {@link Index#createIndex(String)} to include the {@link #alias} to the index creation
-     * request.
+     * Overrides {@link Index#createIndex(String)} to create a Data Stream instead.
      *
-     * @param index Name of the index to create.
-     * @see Alias
+     * @param index Name of the data stream to create.
      */
     @Override
     public void createIndex(String index) {
         try {
-            if (!this.indexExists(index)) {
-                CreateIndexRequest request =
-                        new CreateIndexRequest(index).alias(new Alias(this.alias).writeIndex(true));
-                CreateIndexResponse createIndexResponse =
-                        this.client
-                                .admin()
-                                .indices()
-                                .create(request)
-                                .actionGet(PluginSettings.getTimeout(this.clusterService.getSettings()));
-                log.info(
-                        "Index created successfully: {} {}",
-                        createIndexResponse.index(),
-                        createIndexResponse.isAcknowledged());
-            }
+            this.createDataStream(index);
         } catch (ResourceAlreadyExistsException e) {
-            log.info("Index {} already exists. Skipping.", index);
-        } catch (
-                Exception
-                        e) { // TimeoutException may be raised by actionGet(), but we cannot catch that one.
-            // Exit condition. Re-attempt to create the index also failed. Original exception is rethrown.
+            log.info("Data stream {} already exists. Skipping.", index);
+        } catch (Exception e) {
+            // TimeoutException may be raised by actionGet(), but we cannot catch that one.
+            // Exit condition. Re-attempt to create the data stream also failed. Original exception is
+            // rethrown.
             if (!this.retry_index_creation) {
-                log.error("Initialization of index [{}] finally failed. The node will shut down.", index);
+                log.error(
+                        "Initialization of data stream [{}] finally failed. The node will shut down.", index);
                 throw e;
             }
-            log.warn("Operation to create the index [{}] timed out. Retrying...", index);
+            log.warn("Operation to create the data stream [{}] timed out. Retrying...", index);
             this.retry_index_creation = false;
-            this.indexUtils.sleep(PluginSettings.getBackoff(this.clusterService.getSettings()));
+            this.sleep(PluginSettings.getBackoff(this.clusterService.getSettings()));
             this.createIndex(index);
         }
+    }
+
+    /**
+     * Creates a Data Stream.
+     *
+     * @param name name of the data stream to create.
+     */
+    public void createDataStream(String name) {
+        CreateDataStreamAction.Request request = new CreateDataStreamAction.Request(name);
+
+        AcknowledgedResponse response =
+                this.client
+                        .admin()
+                        .indices()
+                        .createDataStream(request)
+                        .actionGet(PluginSettings.getTimeout(this.clusterService.getSettings()));
+
+        log.info("Data Stream created successfully: {} {}", name, response.isAcknowledged());
     }
 }